Split useConversationScroll into its own hook
