+++ app/src/renderer/views/Inbox/MainContent/Conversation.tsx
@@ -5,19 +5,12 @@ import NewMessagesDivider from "./Conversation/NewMessagesDivider";
 import EmptyConversation from "./EmptyConversation";
 import { Form, Input, Button } from "antd";
 import { useTranslation } from "react-i18next";
-import { useEffect, useRef, memo, useMemo, useCallback, useState } from "react";
-import { useAppDispatch, useAppSelector } from "../../../hooks";
-import {
-  fetchSources,
-  initializeConversationIndicator,
-  markConversationLastSeen,
-  selectConversationLastSeen,
-} from "../../../features/sources/sourcesSlice";
+import { memo, useMemo, useCallback, useState } from "react";
+import { useAppDispatch } from "../../../hooks";
+import { fetchSources } from "../../../features/sources/sourcesSlice";
 import { fetchConversation } from "../../../features/conversation/conversationSlice";
+import useConversationScroll from "./Conversation/useConversationScroll";
 import "./Conversation.css";
-
-const NEW_MESSAGE_SCROLL_OFFSET = 150;
-const NEW_MESSAGE_SEEN_THRESHOLD = 12;
 
 interface ConversationProps {
   sourceWithItems: SourceWithItems | null;
@@ -29,275 +22,16 @@ const Conversation = memo(function Conversation({
   const { t } = useTranslation("MainContent");
   const dispatch = useAppDispatch();
   const [form] = Form.useForm();
-  const scrollContainerRef = useRef<HTMLDivElement>(null);
   const [messageValue, setMessageValue] = useState("");
-  const [dividerItemUuid, setDividerItemUuid] = useState<string | null>(null);
-  const sourceUuidRef = useRef<string | null>(null);
-  const itemCountRef = useRef<number>(0);
-  const dividerUuidRef = useRef<string | null>(null);
-  const pendingScrollTargetRef = useRef<"divider" | "bottom" | null>(null);
-  const dividerRef = useRef<HTMLDivElement | null>(null);
-  const [_isAutoScrolling, setIsAutoScrollingState] = useState(false);
-  const isAutoScrollingValueRef = useRef(false);
-  const setIsAutoScrolling = useCallback(
-    (nextValue: boolean) => {
-      isAutoScrollingValueRef.current = nextValue;
-      setIsAutoScrollingState(nextValue);
-    },
-    [setIsAutoScrollingState],
-  );
-  const activeSourceUuid = sourceWithItems?.uuid ?? null;
-  const journalistUUID = useAppSelector(
-    (state) => state.session.authData?.journalistUUID ?? null,
-  );
-  const lastSeenInteractionCount = useAppSelector((state) =>
-    activeSourceUuid
-      ? selectConversationLastSeen(state, activeSourceUuid)
-      : undefined,
-  );
-
-  const items = useMemo(
-    () => (sourceWithItems ? sourceWithItems.items : []),
-    [sourceWithItems],
-  );
-  const itemCount = items.length;
-  const hasItems = itemCount > 0;
-  const latestInteractionCount = useMemo(() => {
-    return sourceWithItems?.items.at(-1)?.data.interaction_count ?? null;
-  }, [sourceWithItems?.items]);
-
-  // Load last-seen state from whatever the database already considered "seen"
-  // so the divider appears immediately on cold start, before any new scroll events
-  const historicalLastSeenInteractionCount = useMemo(() => {
-    if (!sourceWithItems) {
-      return null;
-    }
-
-    let maxSeen: number | null = null;
-    for (const item of sourceWithItems.items) {
-      const interaction = item.data.interaction_count;
-      if (interaction == null) {
-        continue;
-      }
-
-      let hasBeenSeen = false;
-      if (journalistUUID) {
-        hasBeenSeen = item.data.seen_by.includes(journalistUUID);
-      } else if (item.data.kind !== "reply") {
-        hasBeenSeen = item.data.is_read;
-      } else {
-        hasBeenSeen = true;
-      }
-
-      if (!hasBeenSeen) {
-        continue;
-      }
-
-      if (maxSeen === null || interaction > maxSeen) {
-        maxSeen = interaction;
-      }
-    }
-
-    return maxSeen;
-  }, [journalistUUID, sourceWithItems]);
-
-  const initialLastSeenInteractionCount = useMemo(() => {
-    if (historicalLastSeenInteractionCount !== null) {
-      return historicalLastSeenInteractionCount;
-    }
-    return latestInteractionCount;
-  }, [historicalLastSeenInteractionCount, latestInteractionCount]);
-
-  useEffect(() => {
-    if (!sourceWithItems || lastSeenInteractionCount !== undefined) {
-      return;
-    }
-    dispatch(
-      initializeConversationIndicator({
-        sourceUuid: sourceWithItems.uuid,
-        lastSeenInteractionCount: initialLastSeenInteractionCount,
-      }),
-    );
-  }, [
-    dispatch,
-    initialLastSeenInteractionCount,
-    lastSeenInteractionCount,
-    sourceWithItems,
-  ]);
-
-  useEffect(() => {
-    if (!sourceWithItems) {
-      setDividerItemUuid(null);
-      return;
-    }
-
-    if (lastSeenInteractionCount === undefined || items.length === 0) {
-      setDividerItemUuid(null);
-      return;
-    }
-
-    const threshold =
-      lastSeenInteractionCount === null ? -Infinity : lastSeenInteractionCount;
-    const firstNewItem = items.find((item) => {
-      if (item.data.kind === "reply") {
-        return false;
-      }
-      const interaction = item.data.interaction_count ?? null;
-      if (interaction === null) {
-        return false;
-      }
-      return interaction > threshold;
-    });
-
-    setDividerItemUuid(firstNewItem ? firstNewItem.uuid : null);
-  }, [items, lastSeenInteractionCount, sourceWithItems]);
-
-  // Track conversation transitions and detect when scroll adjustments are needed.
-  useEffect(() => {
-    if (!sourceWithItems) {
-      dividerUuidRef.current = null;
-      sourceUuidRef.current = null;
-      itemCountRef.current = 0;
-      pendingScrollTargetRef.current = null;
-      return;
-    }
-
-    const prevSourceUuid = sourceUuidRef.current;
-    const prevCount = itemCountRef.current;
-
-    if (prevSourceUuid !== sourceWithItems.uuid) {
-      pendingScrollTargetRef.current = dividerItemUuid ? "divider" : "bottom";
-    } else if (dividerItemUuid && dividerItemUuid !== dividerUuidRef.current) {
-      pendingScrollTargetRef.current = "divider";
-    } else if (itemCount > prevCount && !dividerItemUuid) {
-      pendingScrollTargetRef.current = "bottom";
-    }
-
-    sourceUuidRef.current = sourceWithItems.uuid;
-    itemCountRef.current = itemCount;
-    dividerUuidRef.current = dividerItemUuid;
-  }, [dividerItemUuid, itemCount, sourceWithItems]);
-
-  // Jump to just above the divider so the top of the viewport lands slightly before the new items
-  const scrollToDivider = useCallback(() => {
-    const scrollEl = scrollContainerRef.current;
-    const dividerEl = dividerRef.current;
-    if (!scrollEl || !dividerEl) {
-      return false;
-    }
-    const target = Math.max(dividerEl.offsetTop - NEW_MESSAGE_SCROLL_OFFSET, 0);
-    scrollEl.scrollTop = target;
-    return true;
-  }, []);
-
-  // Default behavior for initial loads or when there are no unseen items
-  const scrollToBottom = useCallback(() => {
-    const scrollEl = scrollContainerRef.current;
-    if (!scrollEl) {
-      return false;
-    }
-    scrollEl.scrollTop = scrollEl.scrollHeight;
-    return true;
-  }, []);
-
-  const scheduleAutoScrollReset = useCallback(() => {
-    const reset = () => setIsAutoScrolling(false);
-
-    if (typeof window !== "undefined" && window.requestAnimationFrame) {
-      window.requestAnimationFrame(reset);
-    } else {
-      setTimeout(reset, 0);
-    }
-  }, [setIsAutoScrolling]);
-
-  const acknowledgeNewMessages = useCallback(() => {
-    if (!sourceWithItems) {
-      return;
-    }
-    if (latestInteractionCount === null) {
-      return;
-    }
-    dispatch(
-      markConversationLastSeen({
-        sourceUuid: sourceWithItems.uuid,
-        interactionCount: latestInteractionCount,
-      }),
-    );
-  }, [dispatch, latestInteractionCount, sourceWithItems]);
-
-  // Execute pending scroll actions once the DOM metrics are available
-  useEffect(() => {
-    const target = pendingScrollTargetRef.current;
-    if (!target || !hasItems) {
-      return;
-    }
-
-    setIsAutoScrolling(true);
-
-    if (target === "divider" && dividerItemUuid && scrollToDivider()) {
-      pendingScrollTargetRef.current = null;
-      scheduleAutoScrollReset();
-      return;
-    }
-
-    if (target === "bottom" && scrollToBottom()) {
-      pendingScrollTargetRef.current = null;
-      scheduleAutoScrollReset();
-      return;
-    }
-
-    setIsAutoScrolling(false);
-  }, [
+  const {
+    acknowledgeNewMessages,
     dividerItemUuid,
+    dividerRef,
     hasItems,
-    itemCount,
-    scheduleAutoScrollReset,
-    setIsAutoScrolling,
-    scrollToBottom,
-    scrollToDivider,
-  ]);
-
-  // Clear the divider once the user reaches the end of the list
-  useEffect(() => {
-    const scrollEl = scrollContainerRef.current;
-    if (!scrollEl || !dividerItemUuid) {
-      return;
-    }
-
-    const handleScroll = () => {
-      if (isAutoScrollingValueRef.current) {
-        return;
-      }
-      const distanceToBottom =
-        scrollEl.scrollHeight - (scrollEl.scrollTop + scrollEl.clientHeight);
-      if (distanceToBottom <= NEW_MESSAGE_SEEN_THRESHOLD) {
-        acknowledgeNewMessages();
-      }
-    };
-
-    scrollEl.addEventListener("scroll", handleScroll);
-    return () => {
-      scrollEl.removeEventListener("scroll", handleScroll);
-    };
-  }, [acknowledgeNewMessages, dividerItemUuid]);
-
-  const { oldItems, newItems } = useMemo(() => {
-    if (!dividerItemUuid) {
-      return { oldItems: items, newItems: [] as typeof items };
-    }
-
-    const dividerIndex = items.findIndex(
-      (item) => item.uuid === dividerItemUuid,
-    );
-    if (dividerIndex === -1) {
-      return { oldItems: items, newItems: [] as typeof items };
-    }
-
-    return {
-      oldItems: items.slice(0, dividerIndex),
-      newItems: items.slice(dividerIndex),
-    };
-  }, [dividerItemUuid, items]);
+    newItems,
+    oldItems,
+    scrollContainerRef,
+  } = useConversationScroll(sourceWithItems);
 
   const designation = useMemo(
     () => sourceWithItems?.data.journalist_designation,
